calculator app completed

+++ basic_calculator.py
@@ -1,4 +1,4 @@
 num1 = input('Enter a number: ')
 num2 = input('Enter another number: ')
-result = num1 + num2
-print(int(result))
+result = float(num1) + float(num2)
+print(result)
